+++ Agora-Video-With-FaceUnity-iOS/CapturerAndRender/Render/OpenGL/NV12TextureCache.m
@@ -46,8 +46,8 @@
         pixelBuffer:(CVPixelBufferRef)pixelBuffer
          planeIndex:(int)planeIndex
         pixelFormat:(GLenum)pixelFormat {
-    const int width = CVPixelBufferGetWidthOfPlane(pixelBuffer, planeIndex);
-    const int height = CVPixelBufferGetHeightOfPlane(pixelBuffer, planeIndex);
+    const int width = (int)CVPixelBufferGetWidthOfPlane(pixelBuffer, planeIndex);
+    const int height = (int)CVPixelBufferGetHeightOfPlane(pixelBuffer, planeIndex);
     
     if (*textureOut) {
         CFRelease(*textureOut);
